Added webpack to allow support for CommonJS import modules, restructured the JS code
