+++ src/data.js
@@ -1,0 +1,64 @@
+// Typed data
+const Data = {
+  buttonAccept: {
+    text: "Accept optional cookies",
+    url: "#",
+    id: "accept_optional_cookies",
+    class: "button",
+    tabIndex: 1,
+  },
+  buttonReject: {
+    text: "Reject optional cookies",
+    url: "#",
+    id: "reject_optional_cookies",
+    class: "button",
+    tabIndex: 1,
+  },
+  hideThisMessage: {
+    text: "Hide this message",
+    url: "#",
+    id: "hide_this_message",
+    class: "button",
+    tabIndex: 1,
+  },
+  buttonPreferences: {
+    id: "#btn_preferences",
+  },
+  bannerParagraph: {
+    id: ".cookie-p",
+  },
+  bannerHeadline: {
+    id: ".cookie_head",
+  },
+  bannerWrapper: {
+    id: "#ds-cookie-consent-banner",
+  },
+  cookies: {
+    cookieOne: "dontShowCookieNotice",
+    cookieTwo: "cookies_policy",
+  },
+  formWrapper: {
+    id: "#ds-cookie-consent-form",
+  },
+  acceptMessageAfterInteraction: {
+    text:
+      "You have accepted optional cookies. You can change your cookie settings on the <a href='/latin/legal/cookies/'>Cookies</a> page.",
+    ariaLabel: "Cookie consent confirmation message",
+  },
+  rejectMessageAfterInteraction: {
+    text:
+      "You have rejected optional cookies. You can change your cookie settings on the <a href='/latin/legal/cookies/'>Cookies</a> page.",
+    ariaLabel: "Cookie consent confirmation message",
+  },
+  oldCookieBannerWrapper: {
+    class: ".cookieNotice",
+  },
+  cookiesToRemove: {
+    one: "_ga",
+    two: "_gid",
+    three: "_gat_UA-2827241-1",
+    four: "_gat_UA-2827241-22",
+  },
+};
+
+export default Data;
