+++ src/ds-cookie-consent.ts
@@ -1,186 +1,5 @@
-// Typed data
-const Data = {
-  buttonAccept: {
-    text: "Accept optional cookies",
-    url: "#",
-    id: "accept_optional_cookies",
-    class: "button",
-    tabIndex: 1,
-  },
-  buttonReject: {
-    text: "Reject optional cookies",
-    url: "#",
-    id: "reject_optional_cookies",
-    class: "button",
-    tabIndex: 1,
-  },
-  hideThisMessage: {
-    text: "Hide this message",
-    url: "#",
-    id: "hide_this_message",
-    class: "button",
-    tabIndex: 1,
-  },
-  buttonPreferences: {
-    id: "#btn_preferences",
-  },
-  bannerParagraph: {
-    id: ".cookie-p",
-  },
-  bannerHeadline: {
-    id: ".cookie_head",
-  },
-  bannerWrapper: {
-    id: "#ds-cookie-consent-banner",
-  },
-  cookies: {
-    cookieOne: "dontShowCookieNotice",
-    cookieTwo: "cookies_policy",
-  },
-  formWrapper: {
-    id: "#ds-cookie-consent-form",
-  },
-  acceptMessageAfterInteraction: {
-    text:
-      "You can change your cookie settings on the <a href='/latin/legal/cookies/'>Cookies</a> page.",
-    ariaLabel: "Cookie consent confirmation message",
-  },
-  rejectMessageAfterInteraction: {
-    text:
-      "You have rejected optional cookies. You can change your cookie settings on the <a href='/latin/legal/cookies/'>Cookies</a> page.",
-    ariaLabel: "Cookie consent confirmation message",
-  },
-  oldCookieBannerWrapper: {
-    class: ".cookieNotice",
-  },
-  cookiesToRemove: {
-    one: "_ga",
-    two: "_gid",
-    three: "_gat_UA-2827241-1",
-  },
-};
-
-interface ICookieConsent {
-  globalCookieName?: string;
-  bannerWrapper?: string;
-  innerElemWrapper?: string;
-  createButton(
-    text: string,
-    id: string,
-    className: string,
-    tabindex?: number
-  ): void;
-  createLink(
-    text: string,
-    url: string,
-    id: string,
-    className: string,
-    tabindex?: number
-  ): void;
-  setCookie(
-    name: string,
-    value: string,
-    options: { secure?: boolean; "max-age": number; domain?: string }
-  ): void;
-  checkCookie(name: string): boolean;
-  deleteCookie(...cname: string[]): void;
-}
-
-// Business logic
-const dsCookieConsentBannerAPI: ICookieConsent = ((): any => {
-  // Delete cookie
-  const deleteCookie = (...cname) => {
-    let cookies = document.cookie.split(";");
-    for (let i = 0; i < cookies.length; i++) {
-      let cookie = cookies[i];
-      let eqPos = cookie.indexOf("=");
-      let name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie;
-
-      cname.forEach((c) => {
-        if (name.trim() === c) {
-          document.cookie = `${name}=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/;domain="`;
-          /*below line is to delete the google analytics cookies. they are set with the domain*/
-          document.cookie = `${name}=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/;domain=${location.hostname.replace(
-            /^www\./i,
-            ""
-          )}`;
-        }
-      });
-    }
-  };
-
-  // Set cookie
-  const setCookie = (name, value, options) => {
-    options = {
-      path: "/",
-      ...options,
-    };
-
-    if (options.expires instanceof Date) {
-      options.expires = options.expires.toUTCString();
-    }
-
-    let updatedCookie =
-      encodeURIComponent(name) + "=" + encodeURIComponent(value);
-
-    for (let optionKey in options) {
-      updatedCookie += "; " + optionKey;
-      let optionValue = options[optionKey];
-      if (optionValue !== true) {
-        updatedCookie += "=" + optionValue;
-      }
-    }
-
-    document.cookie = updatedCookie;
-  };
-
-  // Check if cookie exists
-  const checkCookie = (name) => {
-    return -1 !== document.cookie.indexOf(name);
-  };
-
-  // Create link element inside the banner
-  const createLink = (text, url, id, className, tabindex) => {
-    const getInnerElem = document.querySelector(Data.buttonPreferences.id);
-    const createButtonLink = document.createElement("a");
-    const linkText = document.createTextNode(text);
-
-    if (getInnerElem) {
-      const parentElement = getInnerElem.parentNode;
-      createButtonLink.appendChild(linkText);
-      createButtonLink.href = url;
-      createButtonLink.className = className;
-      createButtonLink.id = id;
-      createButtonLink.tabIndex = tabindex;
-      parentElement.insertBefore(createButtonLink, getInnerElem);
-    }
-  };
-
-  // Create link element inside the banner
-  const createButton = (text, id, className, tabindex) => {
-    const getInnerElem = document.querySelector(Data.buttonPreferences.id);
-    const createButtonLink = document.createElement("button");
-    const linkText = document.createTextNode(text);
-
-    if (getInnerElem) {
-      const parentElement = getInnerElem.parentNode;
-      createButtonLink.appendChild(linkText);
-      createButtonLink.className = className;
-      createButtonLink.id = id;
-      createButtonLink.tabIndex = tabindex;
-      parentElement.insertBefore(createButtonLink, getInnerElem);
-    }
-  };
-
-  // Revealing public API
-  return {
-    createButton,
-    createLink,
-    setCookie,
-    checkCookie,
-    deleteCookie,
-  };
-})();
+import Data from "./data";
+import dsCookieConsentBannerAPI from "./api/dsCookieConsentBannerAPI";
 
 const getBannerElement = document.querySelector(Data.bannerWrapper.id);
 const getCookieForm = document.querySelector(Data.formWrapper.id);
